Apply close multi/poly geoms for line annotations as well
