+++ src/mbgl/annotation/shape_annotation_impl.hpp
@@ -3,6 +3,7 @@
 #include <mapbox/geojsonvt.hpp>
 
 #include <mbgl/annotation/annotation.hpp>
+#include <mbgl/util/geometry.hpp>
 
 #include <string>
 #include <memory>
@@ -30,6 +31,35 @@ public:
     const uint8_t maxZoom;
     const std::string layerID;
     std::unique_ptr<mapbox::geojsonvt::GeoJSONVT> shapeTiler;
+};
+
+struct CloseShapeAnnotation {
+    ShapeAnnotationGeometry operator()(const mbgl::LineString<double> &geom) const {
+        return geom;
+    }
+    ShapeAnnotationGeometry operator()(const mbgl::MultiLineString<double> &geom) const {
+        return geom;
+    }
+    ShapeAnnotationGeometry operator()(const mbgl::Polygon<double> &geom) const {
+        mbgl::Polygon<double> closed = geom;
+        for (auto &ring : closed) {
+            if (!ring.empty() && ring.front() != ring.back()) {
+                ring.emplace_back(ring.front());
+            }
+        }
+        return closed;
+    }
+    ShapeAnnotationGeometry operator()(const mbgl::MultiPolygon<double> &geom) const {
+        mbgl::MultiPolygon<double> closed = geom;
+        for (auto &polygon : closed) {
+            for (auto &ring : polygon) {
+                if (!ring.empty() && ring.front() != ring.back()) {
+                    ring.emplace_back(ring.front());
+                }
+            }
+        }
+        return closed;
+    }
 };
 
 } // namespace mbgl
